+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/Autos/BlueVisionAuto.java
@@ -26,21 +26,16 @@ public class BlueVisionAuto extends ULLinearOpMode {
         ArrayList<PathPoint> path = new ArrayList<>();
 
         // fill the path with the correct points, dependent on the number of rings detected
-        if(robot.detectorAuto.ringCount == 0) {
+        if(robot.detector.ringCount == 1) {
             path.clear();
             odom.setStartPosition(53, 8);
-            path.add(new PathPoint(53, 84, 0.7, 4, 5));
-            path.add(new PathPoint(12, 84, 0.7, 4, 5));
-//        } else if(robot.detectorAuto.ringCount == 1) {
-//            path.clear();
-//            odom.setStartPosition(53, 8);
-//            path.add(new PathPoint(53, 108, 0.7, 2, 5));
-//            path.add(new PathPoint(34, 108, 0.7, 2, 5));
-//        } else if(robot.detectorAuto.ringCount == 4) {
-//            path.clear();
-//            odom.setStartPosition(53, 8);
-//            path.add(new PathPoint(53, 132, 0.7, 2, 5));
-//            path.add(new PathPoint(10, 132, 0.7, 2, 5));
+            path.add(new PathPoint(53, 108, 0.7, 4, 5));
+            path.add(new PathPoint(34, 108, 0.7, 4, 5));
+        } else if(robot.detector.ringCount == 4) {
+            path.clear();
+            odom.setStartPosition(53, 8);
+            path.add(new PathPoint(53, 132, 0.7, 4, 5));
+            path.add(new PathPoint(10, 132, 0.7, 4, 5));
         } else {
             path.clear();
             odom.setStartPosition(53, 8);
@@ -48,11 +43,7 @@ public class BlueVisionAuto extends ULLinearOpMode {
             path.add(new PathPoint(12, 84, 0.7, 4, 5));
         }
 
-        telemetry.addData("Top Rectangle Raw Value", (int) robot.detectorAuto.TopMatRaw);
-        telemetry.addData("Top Rectangle Percent", Math.round(robot.detectorAuto.TopMatValue * 100) + "%");
-        telemetry.addData("Bottom Rectangle Raw Value", (int) robot.detectorAuto.BottomMatRaw);
-        telemetry.addData("Bottom Rectangle Percent", Math.round(robot.detectorAuto.BottomMatValue * 100) + "%");
-        telemetry.addData("Number of Rings Detected", robot.detectorAuto.ringCount);
+        telemetry.addData("Number of Rings Detected", robot.detector.ringCount);
         telemetry.update();
 
         // follow the path designated earlier in the program (only if the path list was filled)
@@ -64,4 +55,5 @@ public class BlueVisionAuto extends ULLinearOpMode {
         odom.stopUpdateThread();
 
     }
+
 }
